# backend/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Artifex Gallery API",
    "version": "1.0.0",
    "description": "AI-powered image and video gallery with auto-tagging, NSFW detection, and federation support."
  },
  "servers": [{ "url": "/api" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/collections": {
      "get": { "tags": ["Collections"], "summary": "List collections", "responses": { "200": { "description": "Collection list" } } },
      "post": {
        "tags": ["Collections"], "summary": "Create collection", "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "description": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/collections/{id}": {
      "get": { "tags": ["Collections"], "summary": "Get collection detail", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Collection with images" } } },
      "put": { "tags": ["Collections"], "summary": "Update collection", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Updated" } } },
      "delete": { "tags": ["Collections"], "summary": "Delete collection", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Deleted" } } }
    },
    "/collections/{id}/images": {
      "post": { "tags": ["Collections"], "summary": "Add images to collection", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Added" } } }
    },
    "/collections/{id}/reorder": {
      "put": { "tags": ["Collections"], "summary": "Reorder collection images", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Reordered" } } }
    },
...
  }
}